+++ admin/src/main/java/kernel360/ckt/admin/application/CompanyService.java
@@ -3,8 +3,8 @@ package kernel360.ckt.admin.application;
 import kernel360.ckt.admin.ui.dto.request.CompanyCreateRequest;
 import kernel360.ckt.admin.infra.repository.CompanyRepositoryImpl;
 import kernel360.ckt.admin.ui.dto.request.CompanyUpdateRequest;
-import kernel360.ckt.core.common.CompanyErrorCode;
-import kernel360.ckt.core.common.CustomException;
+import kernel360.ckt.core.common.error.CompanyErrorCode;
+import kernel360.ckt.core.common.exception.CustomException;
 import kernel360.ckt.core.domain.entity.CompanyEntity;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
