+++ core/src/main/java/kernel360/ckt/core/common/error/CompanyErrorCode.java
@@ -1,4 +1,4 @@
-package kernel360.ckt.core.common;
+package kernel360.ckt.core.common.error;
 
 import lombok.Getter;
