+++ core/src/main/java/kernel360/ckt/core/common/error/ErrorCode.java
@@ -1,4 +1,4 @@
-package kernel360.ckt.core.common;
+package kernel360.ckt.core.common.error;
 
 public interface ErrorCode {
     String getCode();
